Created generalized IntHasher for optimized hashing of type IDs and router component
