+++ src/utils/hashing.rs
@@ -1,8 +1,10 @@
 //! Hashing utility for hashing and verifying passwords
 
+use std::hash::Hasher;
+
 use argon2::{
-    Argon2,
-    password_hash::{self, PasswordVerifier, rand_core::OsRng, SaltString}, PasswordHash, PasswordHasher,
+    password_hash::{self, rand_core::OsRng, PasswordVerifier, SaltString},
+    Argon2, PasswordHash, PasswordHasher,
 };
 
 /// Hashes the provided password using the Argon2 algorithm returning
@@ -29,4 +31,34 @@ pub fn verify_password(password: &str, hash: &str) -> bool {
     };
     let argon2 = Argon2::default();
     argon2.verify_password(password.as_bytes(), &hash).is_ok()
+}
+
+/// Hasher implementation that directly uses an integer value
+/// instead of any specific hashing algorithm
+///
+/// Only implements hashing for [u32] and [u64]
+///
+/// Used for hashing packet component paths and type Ids
+#[derive(Default)]
+pub struct IntHasher(u64);
+
+impl Hasher for IntHasher {
+    fn write(&mut self, _: &[u8]) {
+        panic!("Attempted to use int hasher to hash bytes")
+    }
+
+    #[inline]
+    fn write_u64(&mut self, id: u64) {
+        self.0 = id;
+    }
+
+    #[inline]
+    fn write_u32(&mut self, i: u32) {
+        self.0 = i as u64;
+    }
+
+    #[inline]
+    fn finish(&self) -> u64 {
+        self.0
+    }
 }
